Changes run.sh to improve cleanup and allow for test flag

+++ run.sh
@@ -1,5 +1,24 @@
 #!/bin/bash
 
-docker rm -vf $(docker ps -aq)
+#Run with flag -t to use test dataset
 
-docker-compose up --build
+if [[ -n $(docker ps -aq) ]]; then
+    echo "Deleting previous containers"
+    docker rm -vf $(docker ps -aq)
+fi
+
+DB="solr/recipes.json"
+
+if [[ $1 == "-t" ]]; then
+    echo "Using test dataset. NOT FIT FOR PRODUCTION!"
+    DB="solr/recipes_test.json"
+fi
+
+docker-compose build --build-arg DB=$DB
+
+if [[ -n $(docker image ls -f dangling=true) ]]; then
+    echo "Removing dangling images"
+    docker rmi -f $(docker image ls -f dangling=true)
+fi
+
+docker-compose up
